Move torch related influence tests to torch test subpackage
